requirements.txt
gpxpy>=1.6
pyproj>=3.6
numpy>=1.26
scipy>=1.12
requests>=2.31
rasterio>=1.3
numpy-stl>=3.1
